// webportal/src/app/enums/sieve-operators.ts
export enum SieveOperators {
  Equals = '==',
  NotEquals = '!=',
  GreaterThanEqualTo = '>=',
  LessThanEqualTo = '<=',
  GreaterThan = '>',
  LessThan = '<',
  Contains = '@=',
  StartsWith = '_=',
  NotContains = '!@=',
  NotStartsWith = '!_=',
  CI_Contains = '@=*',
  CI_StartsWith = '_=*',
  CI_Equals = '==*',
  CI_NotEquals = '!=*',
  CI_NotContains = '!@=*',
  CI_NotStartsWith = '!_=*',
}